+++ biceps/src/main/java/org/somda/sdc/biceps/consumer/access/RemoteMdibAccessImpl.java
@@ -122,6 +122,11 @@ public class RemoteMdibAccessImpl implements RemoteMdibAccess {
     }
 
     @Override
+    public void unregisterAllObservers() {
+        eventDistributor.unregisterAllObservers();
+    }
+
+    @Override
     public MdibVersion getMdibVersion() {
         try (ReadTransaction transaction = startTransaction()) {
             return transaction.getMdibVersion();